+++ src/azure-cli-core/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 2.0.29
 ++++++
 * Support Autorest 3.0 based SDKs
+* Support mechanism for a command module to suppress the loading of particular extensions.
 
 2.0.28
 ++++++